HHH-9568 : Move test case from CascadeTest to OneToOneOrphanTest
(cherry picked from commit 083e8cf52ef18a7eaa965612a1829314fe0ff66a)
